fixed: change getTweet to getReplies
