+++ controllers/replyController.js
@@ -1,16 +1,47 @@
 const helpers = require('../_helpers')
-const { User, Reply, Tweet } = require('../models')
+const { User, Reply, Tweet, Like } = require('../models')
 const replyController = {
   getReply: async (req, res, next) => {
     try {
-      const replies = await Reply.findAll({
-        where: { TweetId: req.params.tweet_id },
-        include: [User, { model: Tweet, include: User }]
+      let tweet = await Tweet.findByPk(req.params.tweet_id,
+        {
+          include: [
+            User, Like, Reply,
+            { model: Reply, include: User }
+          ],
+          order: [['createdAt', 'DESC']]
+        })
+
+      let users = await User.findAll({
+        include: [
+          { model: User, as: 'Followers' },
+          { model: User, as: 'Followings' }
+        ]
       })
-      //render要修改
-      return res.render('tweet', { replies: replies.toJSON() })
-    }
-    catch (err) {
+
+      const user = await User.findByPk(helpers.getUser(req).id,
+        {
+          raw: true,
+          nest: true
+        })
+
+      const likes = await Like.findAll({
+        where: { UserId: helpers.getUser(req).id }
+      })
+      const likedCount = tweet.Likes.length
+      const repliedCount = tweet.Replies.length
+      const isLiked = likes.map(l => l.TweetId).includes(tweet.id)
+
+
+      users = await users.map(user => ({
+        ...user.toJSON(),
+        followerCount: user.Followers.length,
+        isFollowed: helpers.getUser(req).Followings.map(f => f.id).includes(user.id)
+      }))
+      users = users.sort((a, b) => b.followerCount - a.followerCount)
+        .slice(0, 10)
+      return res.render('tweet', { tweet: tweet.toJSON(), users, user, likedCount, repliedCount, isLiked })
+    } catch (err) {
       next(err)
     }
   },
@@ -35,7 +66,7 @@ const replyController = {
         comment
       })
       req.flash('success_messages', '成功新增回覆！')
-      return res.redirect(`/tweets/${req.params.tweet_id}`)
+      return res.redirect(`/tweets/${req.params.tweet_id}/replies`)
     }
     catch (err) {
       next(err)
