+++ routes/index.js
@@ -17,7 +17,6 @@ router.post('/tweets/:tweet_id/replies', authenticated, replyController.postRepl
 router.post('/tweets/:tweet_id/unlike', authenticated, tweetController.postUnlike)
 router.post('/tweets/:tweet_id/like', authenticated, tweetController.postLike)
 router.get('/tweets/:tweet_id/replies', authenticated, replyController.getReply)
-router.get('/tweets/:tweet_id', authenticated, tweetController.getTweet)
 router.get('/tweets', authenticated, tweetController.getTweets)
 router.post('/tweets', authenticated, tweetController.postTweet)
 router.delete('/followships/:followingId', authenticated, followshipController.removeFollowing)
