+++ views/tweets.hbs
@@ -44,7 +44,7 @@
               <a class="account" href="/users/:{{this.User.id}}">@{{this.User.account}}</a>
               <div class="time">•3小時</div>
             </div>
-            <a href="/tweets/{{this.id}}">
+            <a href="/tweets/{{this.id}}/replies">
               <div class="content">
                 {{this.description}}
               </div>
